Développement des étapes d'authentification terminé

+++ controllers/AccountController.php
@@ -2,6 +2,7 @@
 
 require_once __DIR__ . '/../models/repositories/ClientRepository.php';
 require_once __DIR__ . '/../models/repositories/AccountRepository.php';
+require_once __DIR__ . '/../lib/utils.php';
 
 class AccountController
 {
@@ -16,66 +17,94 @@ class AccountController
 
     public function create()
     {
-        $clients = $this->clientRepository->viewClients();
+        if (isConnected()) {
+            $clients = $this->clientRepository->viewClients();
 
-        require_once __DIR__ . '/../views/accountViews/account-create.php';
+            require_once __DIR__ . '/../views/accountViews/account-create.php';
+
+            exit;
+        }
     }
 
     public function set()
     {
-        $account = new Account();
-        $account->setRib($_POST['rib']);
-        $account->setType($_POST['type']);
-        $account->setSolde($_POST['balance']);
-        $account->setIdClient($_POST['clientNum']);
+        if (isConnected()) {
+            $account = new Account();
+            $account->setRib($_POST['rib']);
+            $account->setType($_POST['type']);
+            $account->setSolde($_POST['balance']);
+            $account->setIdClient($_POST['clientNum']);
 
-        $this->accountRepository->addAccount($account);
+            $this->accountRepository->addAccount($account);
 
-        header('Location: ?action=account-list');
+            header('Location: ?action=account-list');
+                        
+            exit;
+        }
 
     }
 
     public function list()
     {
-        $accounts= $this->accountRepository->viewAccounts();
+        if (isConnected()) {
+            $accounts = $this->accountRepository->viewAccounts();
 
-        require_once __DIR__ . '/../views/accountViews/account-list.php';
+            require_once __DIR__ . '/../views/accountViews/account-list.php';
+                        
+            exit;
+        }
     }
 
     public function show(int $id)
     {
-        $account = $this->accountRepository->viewAccount($id);
-        $client = $this->clientRepository->viewClient($account->getIdClient());
+        if (isConnected()) {
+            $account = $this->accountRepository->viewAccount($id);
+            $client = $this->clientRepository->viewClient($account->getIdClient());
 
-        require_once __DIR__ . '/../views/accountViews/account-show.php';
+            require_once __DIR__ . '/../views/accountViews/account-show.php';
+                        
+            exit;
+        }
     }
 
     public function edit(int $id)
     {
-        $account = $this->accountRepository->viewAccount($id);
+        if (isConnected()) {
+            $account = $this->accountRepository->viewAccount($id);
 
-        require_once __DIR__ . '/../views/accountViews/account-edit.php';
+            require_once __DIR__ . '/../views/accountViews/account-edit.php';
+                        
+            exit;
+        }
     }
 
     public function update()
     {
-        $account = new Account();
-        $account->setId($_POST['id']);
-        $account->setRib($_POST['rib']);
-        $account->setType($_POST['type']);
-        $account->setSolde($_POST['balance']);
-        $account->setIdClient($_POST['clientNum']);
+        if (isConnected()) {
+            $account = new Account();
+            $account->setId($_POST['id']);
+            $account->setRib($_POST['rib']);
+            $account->setType($_POST['type']);
+            $account->setSolde($_POST['balance']);
+            $account->setIdClient($_POST['clientNum']);
 
-        $this->accountRepository->updateAccount($account);
+            $this->accountRepository->updateAccount($account);
 
-        header('Location: ?action=account-list');
+            header('Location: ?action=account-list');
+                        
+            exit;
+        }
     }
 
     public function delete(int $id)
     {
-        $this->accountRepository->deleteAccount($id);
+        if (isConnected()) {
+            $this->accountRepository->deleteAccount($id);
 
-        header('Location: ?action=account-list');
+            header('Location: ?action=account-list');
+                        
+            exit;
+        }
     }
