+++ controllers/ClientController.php
@@ -1,6 +1,7 @@
 <?php
 
 require_once __DIR__ . '/../models/repositories/ClientRepository.php';
+require_once __DIR__ . '/../lib/utils.php';
 
 class ClientController
 {
@@ -13,65 +14,86 @@ class ClientController
 
     public function create()
     {
-        require_once __DIR__ . '/../views/clientViews/client-create.php';
+        if (isConnected()) {
+            require_once __DIR__ . '/../views/clientViews/client-create.php';
+            exit;
+        }
     }
 
     public function set()
     {
-        $client = new Client();
-        $client->setNom($_POST['name']);
-        $client->setPrenom($_POST['surname']);
-        $client->setEmail($_POST['email']);
-        $client->setTelephone($_POST['phone']);
-        $client->setAdresse($_POST['location']);
+        if (isConnected()) {
+            $client = new Client();
+            $client->setNom($_POST['name']);
+            $client->setPrenom($_POST['surname']);
+            $client->setEmail($_POST['email']);
+            $client->setTelephone($_POST['phone']);
+            $client->setAdresse($_POST['location']);
 
-        $this->clientRepository->addClient($client);
+            $this->clientRepository->addClient($client);
 
-        header('Location: ?action=client-list');
+            header('Location: ?action=client-list');
+            exit;
+        }
 
     }
 
     public function list()
     {
-        $clients= $this->clientRepository->viewClients();
+        if (isConnected()) {
+            $clients = $this->clientRepository->viewClients();
 
-        require_once __DIR__ . '/../views/clientViews/client-list.php';
+            require_once __DIR__ . '/../views/clientViews/client-list.php';
+            exit;
+        }
     }
 
     public function show(int $id)
     {
-        $client = $this->clientRepository->viewClient($id);
+        if (isConnected()) {
+            $client = $this->clientRepository->viewClient($id);
 
-        require_once __DIR__ . '/../views/clientViews/client-show.php';
+            require_once __DIR__ . '/../views/clientViews/client-show.php';
+            exit;
+        }
     }
 
     public function edit(int $id)
     {
-        $client = $this->clientRepository->viewClient($id);
+        if (isConnected()) {
+            $client = $this->clientRepository->viewClient($id);
 
-        require_once __DIR__ . '/../views/clientViews/client-edit.php';
+            require_once __DIR__ . '/../views/clientViews/client-edit.php';
+            exit;
+        }
     }
 
     public function update()
     {
-        $client = new Client();
-        $client->setId($_POST['id']);
-        $client->setNom($_POST['name']);
-        $client->setPrenom($_POST['surname']);
-        $client->setEmail($_POST['email']);
-        $client->setTelephone($_POST['phone']);
-        $client->setAdresse($_POST['location']);
+        if (isConnected()) {
+            $client = new Client();
+            $client->setId($_POST['id']);
+            $client->setNom($_POST['name']);
+            $client->setPrenom($_POST['surname']);
+            $client->setEmail($_POST['email']);
+            $client->setTelephone($_POST['phone']);
+            $client->setAdresse($_POST['location']);
 
-        $this->clientRepository->updateClient($client);
+            $this->clientRepository->updateClient($client);
 
-        header('Location: ?action=client-list');
+            header('Location: ?action=client-list');
+            exit;
+        }
     }
 
     public function delete(int $id)
     {
-        $this->clientRepository->deleteClient($id);
+        if (isConnected()) {
+            $this->clientRepository->deleteClient($id);
 
-        header('Location: ?action=client-list');
+            header('Location: ?action=client-list');
+            exit;
+        }
     }
